requirements.txt
pdfplumber
gTTS
google-generativeai
